#50 refactor: CurrentPositionMarker の position 反映 useEffect をカスタムフックへ切り出し
